+++ CHANGES.md
@@ -2,7 +2,8 @@
 
 * **BREAKING**: Removed type-level interceptors from ring-adapter, use normal interceptors instead.
 * **BREAKING**: Ring request-parameters are now assoc-in'd (into `:data`) instead of deep-merging. For speed.
-* Handlers can be now be mounted to dispatcher root
+* Handlers can be now be mounted to dispatcher root.
+* Removed `kekkonen.core/simple-coercion`, renamed `multi-coercion` to `coercion`.
 * Support for Context-based urls, thanks to [Wout Neirynck](https://github.com/wneirynck).
 * Data input schemas for apis can be vectors, fixes [#27](https://github.com/metosin/kekkonen/issues/27).
 * Use Pedestal-style interceptors, with `:name`, `:enter`, `:leave` and `:error`